Refs #8587. Moved implementation to specific folder.
