+++ src/cpp/utils/collections/impl/node-sizes/custom/list_node_size_impl.hpp
